Add siteParam: Button text (Accessibility, NoScript); TODO: conditional NOT for header.html
Similar to how it is done on this blog: https://justabloginthepark.com/photos/
Navbar buttons need (optional, toggleable) text next to them, in case of NoScript/incompatible browser blocking the icons, or just plain old ease-of-use for Grandma.
Separate button text for when Archives/Categories/Tags buttons are small, below site title. TODO: Need to do a NOT along the lines of:
{{ if .Site.Params.textButtonsMini NOT .Site.Params.topTags }}